Add capacity to airship CI cloud
The airship jobs are expected to need 1x32GB ram instance + 3x16GB ram
instances each. The goal is to run 4 jobs concurrently which gives us 16
max-instances in the airship pool. We set the main pool to max 10
instances of the smaller 8GB ram instances to act as canaries in the
main pool.

Change-Id: Ie0293d0523f8a9a738015035725ea2e40b152d3d