Publish tarballs artifacts to AFS in parallel
This adds publishing of artifacts to AFS in parallel with the existing
publishing to the static volume.

The additional tasks added to the openstack-artifacts.yaml playbook
have been tested in the project-config-test-artifacts-to-afs test
(results at [1]) and also in an external project
ironic-python-agent-builder (see notes in [2]).  Thus we can have high
confidence it will work.

After this merges, I will manually rsync the old static tarballs data
to the AFS directory (again; it has been sync'd previously) and from
that point on both should be exactly the same.

At that point we can switch the DNS entries for
static.<openstack|opendev>.org (the sites have been created with [3])
to the new server and we will be fully migrated to the AFS volume.  We
can then remove the static publishing.

[1] http://files.openstack.org/project/tarballs.opendev.org/openstack/project-config/afs-copy-test/
[2] https://review.opendev.org/705161
[3] https://review.opendev.org/704913

Change-Id: Ia3a0358249e9ed3d766b1b61535f2f6d67d4eb2d