Stop running py2-based jobs for devstack-gate
The python2-based jobs from integrated-gate are failing now since
Neutron requires py>=3.6. So drop that template from the devstack-gate
project.

The alternative would be to run those jobs against a stable branch
instead master, but that would be non-trivial since devstack-gate is
branchless.

Also the expected remaining lifetime of ds-gate is limited, since it is
a current community goal to convert all legacy jobs to new versions,
which will obsolete the devstack-gate code. With that argument I also
refrain from moving the whole job setup into the project itself, as we
would do with any project that still has a future.

Change-Id: If094fcc31929193eb8278bdbb182e24aa506cca7