check-release-approval: retry once when Gerrit fails
We've had a number of occurrences where Gerrit failed to return a
complete JSON answer. In particular, the ['labels']['CodeReview']['all']
content is not provided, while the query requires o=DETAILED_LABELS.

Since this is a rare occurrence (which could not be reproduced on
manual testing) let's retry once to load the results from Gerrit.

Change-Id: I98d1010e9586d2329137d5d600c66eeb8343fa97