Add publish job for stackviz
stackviz gets published daily to:
https://tarballs.opendev.org/openstack/stackviz/dist/stackviz-latest.tar.gz

The current job periodic-package-stackviz-element is a legacy job,
create a new job publish-openstack-javascript-latest-content that can be
run instead.

The stackviz-element is backed into our images and thus we have to be
extra careful not to break it and test.

The new job is publishing temporarily under a different filename to
check that it's working correctly.

Order of changes:
* Merge this change, so we run both periodic-package-stackviz-element and
  publish-openstack-javascript-latest-content in stackviz.
* Change new job to publish under correct name and remove setup in
  project-config.
* Switch to new job in stackviz repo.

Change-Id: I81ac0975f03756da673ae12e7e11cd5686a0f2d6