Move stackviz publish in-tree
The new job publish-openstack-stackviz-element is working now,
move the trigger in-tree and change the publish location, we do not need
the "-new" in the name anymore.

Content is published at:
https://tarballs.opendev.org/openstack/stackviz/dist/
The tarballs are equal, only difference is filenames of sources.

Mark the job to run only on master, since it publishes always to the
same location and each invocation overrides the file. This is a
safety measure.

Depends-On: https://review.opendev.org/710754
Change-Id: If9488d7c60b58f51a7a0a4bdf0740d16f61c241e