Restore blank nodepool/scripts
I was a little too hasty in I76acbd08acda20c88ff9fd9148e3918b78d6c6c6
removing the scripts/ directory.  It has broken the puppeting of old
hosts.

Restore the directory with a blank file explaining the situation.

Also, we don't need to copy this in the nodepool elements; remove that.

Change-Id: I8b82950237ef69c4941866900cac9bda42f58ca2