nodepool-elements: skip tox install with no pip-and-virtualenv
This is a follow-on to I85438baf5bb31790a56fe5b38327361f0a2398e9.

Skip over this install of tox, which no longer works without the
"pip-and-virtualenv" element define of $DIB_PYTHON_PIP.  We want to
not install globally in the image, but move things like this to
ansible roles in base jobs if required.

Change-Id: Id1571210f0778019c78aec9f38e9f1254c1d68f9