Add OpenInfra Labs IRC channels to bots
The OpenInfra Labs community is utilizing OpenDev to to collaborate
(already have a mailing list and Git repository in OpenDev), and are
interested in having some of OpenDev's bots assist with their IRC
channels as well. Add these channels to the accessbot configuration
first, to pave the way for others which will follow. Also set up
gerritbot to report all changes for their repository namespace in
their primary channel.

Change-Id: Ief1a0ef61bc446f6013fb1ad3968cef2e72e8508