Revert "Move fedora-30 builds to nb01.opendev.org"
This reverts commit 9d7592df75a84cac484457e62dce917af80b092a which was
meant to make nb01.opendev.org the only builder of fedora-30, but did
not quite work as expected.  This replicated what we do on the ARM64
builds; but what I missed was that the ARM builds have exclusive use
of the ARM clouds -- in this case the new nb01.opendev.org thought
that all the other images in the existing clouds were foreign and
tried to delete them.

Take a different approach; replicate the nodepool.yaml file to a
nb01.opendev.org specific one, but pause everything *but* the
fedora-30 builds.  This should make the builder leave everything else
alone, and just try building Fedora, which is what we want for the
initial POC.

Change-Id: I2ccb6aaad5866561fd6ec4118c718ee131c3a5f0