Install tox into a virtualenv on our images
This will install tox into a virtualenv on our images. On our older
images with globally installed tox this can be ignored but as we move to
"plain" images this can be used as an opt in tox executable by jobs.
Jobs can set the tox_executable path for the ensure-tox role.

We don't install it globally to avoid polluting package manager managed
paths.

Change-Id: If5397d731e9fb04431482529aed23cd9fdaecc1d