Verbosely apply Zuul ssh key in nodepool config
We had been relying on the ssh pubkey configured in the nodepool user's
homedir. This being an implicit config caused containerized nodepool
builders to break as they didn't have this config. Switch to explicit
config to avoid this in the future. This is unfortunately quite verbose,
but that will improve with future changes to nodepool.

Change-Id: I67d0c61be9cd05c90a5be937fa4bf77678f89ce5
Story: #2007407
Task: #39072