Use /detail in check-release-approval queries
Gerrit sometimes returns a partial JSON answer, missing
the details that o=DETAILED_LABELS should trigger. This
results in false negatives in check-release-approval tests.

This cannot be reproduced easily. We put in place a retry
but the issue seems to stick on immediate retries.

As an experiment, this change switches the API call from
/changes/ID with o=DETAILED_LABELS&o=DETAILED_ACCOUNTS to
/changes/ID/detail (which includes these two options, amongst
others), to see if that would workaround the Gerrit issue.

We also remove the retry since it does not improve significantly
the situation.

Change-Id: I4de49da1b48f7b87879102a0e18e97168e39406b