Add whitebox-tempest-plugin under QA
Tempest defines its scope as only what is accessible through the
various REST APIs. Some cloud features cannot be properly tested when
using only the REST API. The whitebox-tempest-plugin is a Tempest
plugin whose scope explicitly requires peeking behind the curtain. In
other words, if a feature or behavior can be fully tested using only a
REST API, such a test does not belong in whitebox-tempest-plugin. On
the other hand, if fully testing a feature or behavior requires
accessing the control plane like a human operator or admin would, such
a test belongs in whitebox-tempest plugin.

Now that the spec [1] for whitebox has merged, this patch moves the
whitebox-tempest-plugin from x to openstack (QA).

[1] https://review.opendev.org/#/c/656890/

Change-Id: Ide4bdf1de6a8ce5d1f86ea82a9fbf5b8a8bf437d