Run gerrit image jobs on jeepyb patches
We're building gerrit container images with jeepyb built in. This
means we should trigger new images to be built when we land jeepyb
patches.

Once we have a manage-projects functional test in system-config
we can add that here too.

While we're touching it - remove the pep8 job, we can
set that in the project's .zuul.yaml.

Change-Id: Ib09c74314b705655e70a02fcc07316d0d080d533