Apply gerrit jobs directly to jeepyb
There is currently a zuul bug that is attaching runnability to
project-template definition and so it's seeing that since this
project-template is defined in system-config which is not a config
project that jeepyb can't run the upload jobs. That's not what
we want to happen.

For now, work around it by just expanding the template here. Once
the zuul bug is fixed, we can revert this change.

Change-Id: Ib9c1ec47024a93571b3c1289ba3ab537417e6830