Revert "Revert "Disable github reporting and re-add devel ansible job""
This reverts commit 9e9aeff0fc998bebfe045fc865797b102d4ec8d4.

We were not done debugging.  In fact, we hadn't even started debugging,
so revert the revert of the job removal.  Here's the justification for
the original revert:

We had a job trigger that shouldn't have. Re-add the job so
we can debug why that happened. Because we're debugging, turn of
all github reporting.

Change-Id: Idb16bcf5b581e7dd411506eb69c2324e206f0511