Replace incident channel with opendev-meeting
The openstack-infra-incident channel has traditionally (though
infrequently) been used as a venue for focusing on specific tasks
during emergency incidents, so that conversation doesn't get muddied
by other discussions occurring in the general channel. The OpenDev
collaborators are also in need of a brand-neutral channel where we
can conduct recurring IRC meetings, instead of continuing to rely on
the openstack-meeting channel. The two needs are actually similar,
and also incidents and meeting conversation will tend to not occur
at the same exact time (because if there's an incident underway
we'll be postponing the meeting, or actually meeting around the
incident, depending on how you look at it).

Change-Id: I56663ad0de2db297956c8a9ad36b005b52d20065