vexxhost: move base-jobs to config-project
Inside the VEXXHOST tenant, we have a need to be able to use build
Docker images in many different places.  Therefore, we need the
ability to have secrets inside of a repository which other repos
can just use the jobs for, avoiding the need of encrypting the
Docker credentials for every single repository.

However, due to the current limitation in Zuul, it's not possible
to accomplish this without having a config-project, and by being
a config-project, that provides an elevated set of access.  As
an interim solution until Zuul has the ability to do this without
using a config-project, this change makes the project a config
project however changes the ACLs to include project-config-core.

The rationale was that I (mnaser) is already part of that group
and therefore this wouldn't be providing me any more access to
make changes to config projects.  This would be an interim solution
until we're able to do this natively with Zuul and the ACLs can
return to VEXXHOST.

In this change, we also move opendev/project-config to only load
jobs, secrets and nodesets and to avoid loading the project so we
don't end up reporting to changes to opendev/project-config.

Change-Id: I6baefcae3e23767aeeaa2d572b1a17fd2aa5ebe6