Fedora 31 : remove pip-and-virtualenv
As part of [1] we are removing pip-and-virtualenv.  dib 2.35.0 merged
an accidental breakage for Fedora [2].  We might as well use this as
an opportunity to have this as a "bare" test platform; where we leave
configuration of pip/virtualenv up to jobs.  This will be alongside
the ubuntu-bionic-plain images (see
I85438baf5bb31790a56fe5b38327361f0a2398e9).

[1] https://docs.opendev.org/opendev/infra-specs/latest/specs/cleanup-test-node-python.html
[2] https://review.opendev.org/716795

Change-Id: I56a2495e9c768d744ac39073c42a5a1754612f70