release-approval pipeline: fix zuul-excluding regexp
The release-approval pipeline uses a username filter in the
comment-added trigger so that comments from 'zuul' username would
not trigger it. Unfortunately, the regexp also matched usernames
which contained a dot, like 'rico.lin'.

Fix the regexp so that it matches 'ttx', 'rico.lin', 'zuulfoo' and
'zuul.bar', but not 'zuul'.

See tests for the old and new regexp at:
http://paste.openstack.org/show/791583/

Change-Id: I6834f71b52570310735f202984af8f7f6ce701a1