Introduce job for granular GitHub mirroring
Create a openstack-mirror-on-github job that uses the
openstack-mirroring user on GitHub (and its SSH key) to
individually mirror repositories to the "openstack"
organization on GitHub.

That will allow us to turn off global replication of the
'openstack/* repositories at Gerrit-level.

Enable the new job on openstack/release-test as a test. It
should be able to run concurrently with Gerrit-wide
mirroring as this is an indempotent operation.

Once this merges and is validated, we can apply the job to
all official/active openstack projects, before disabling the
Gerrit-wide replication (see https://review.opendev.org/718478).

Change-Id: Ie7b4f520d6f47d56a71c812dcc06fd5d26da8fe0