Move suse builds to nb04, drop pip-and-virtualenv
Move these images to nb04, the container-based builder.  Pause them on
the other builders.  Due to upstream changes the pip-and-virtualenv
element doesn't work on them any more; remove it.  Work is in progress
to fix our base jobs to operate on platforms such as this, but in the
mean time devstack tests should work.

Change-Id: I0b5ccf6f0b441bdb752b5f066b69edda92f4b37a