[check-approval] Use committer instead of owner
The check-release-approval job currently considers a change being
approved if a release liaison is the change owner, or if a release
liaison approves it in a subsequent vote.

This means that if a release liaison pushes the original change (and
is therefore the change owner), any patchset pushed above the original
one will retain the owner and therefore the approval.

Instead of the change owner, use the committer of the last revision.

Change-Id: I2f7730626a8785b63f88fdf5d507fbd97717b8f3