Revert "Revert "Introduce job for granular GitHub mirroring""
This change inroduces the usage of the parent job which enforces
a project name inside of it to avoid a project from mirroring
maliciously to another one.

It reparents it to the nodeless job inside opendev, enforeces the
setting of target_repository inside the secret so it cannot be modified
and fixes the `user` to point to `git` instead.

This reverts commit 04133b4bccb931b403a22ddcd49737e02907dd2a.

Change-Id: Iac5db122eb41709dffcdae675137723cd49fca8f