nb03 : update to arm64 to inheritance, drop pip-and-virtualenv
This updates the nb03 configuration file to use inheritance, removing
a lot of duplication.

We remove the labels: section which isn't required in the builder
configuration.

We also drop pip-and-virtualenv from these images.  For the x86 images
we're staging things a bit more with "-plain" images, but we don't
have room on the builders to double-build everything, and these are
less likely to be gating.

Change-Id: I1f810c160e5d643f332ffde06792e4b5626a557e