Use zuul deployment keys for yaml2ical
We're currently adding the private key of a jenkins user. But we
don't use jenkins, so this is lame. Use the zuul deploy key for
project-config instead.

This needs to run as project-config so that per-project deploy
keys work. To do that, shift it from being triggered by irc-meetings
to being triggered hourly by project-config.

As a followup, we should probably convert this to just publish to
AFS and serve the content from there.

Depends-On: https://review.opendev.org/721098
Change-Id: I7874ef46a616e8fd68cf8d95afc3928d3440ba51