Add Airship subproject documentation job
The airship/treasuremap project publishes to docs.airshipit.org, but
other airship repositories cannot publish to the same site. This change
adds a separate job that enables other Airship repositories to promote
their documentation to a subdirectory of the main documentation site.
This change also replaces airship/treasuremap with airship/docs as the
main documentation repository.

Change-Id: I703ac705b549f4f4a7bc24d37a63b88035b5e892
Signed-off-by: Drew Walters <andrew.walters@att.com>