Add tripleo-ci-shared-core group
Adds new group tripleo-ci-shared-core group which will include
tripleo-core and be used to grant access to other collaborators to
specific list of repositories which contain shared code.

Once the new group is created, I will add tripleo-core as a subgroup
and the members from the retired openstack-tempest-skiplist group.

The purpose of this group is to ease ACL maintenance by re-using
it among several similar projects.

Change-Id: I52de44bca33d8faea10fa403e53959a194ff2889