Run the zuul service playbook on tenant changes
We're now running Zuul configuration from the service-zuul Ansible
playbook rather that remote-puppet-else.  Use that so that changes
to the Zuul tenant configuration take effect.

Also, make that job soft-depend on manage-projects -- so if we're
adding a new repo to gerrit, we make sure that happens before we
update the Zuul config.

Change-Id: I0bbab2d86cd1d3bd7d9a0469a04868e4069d7479