Remove tumbleweed-plain images
We took a slightly different approach for tumbleweed and dropped
support for the pip-and-virtualenv element for it (and Fedora 31+)
with I7a6a342461d6001c25e55638ba9b7438c28f2519.  So there's no need
for the -plain image now; switch the main image to using the base pip.

This is currently failing to build due to this.

Change-Id: I3333b475ff9d2f2fb96ad5588c6250cba2593e06