Restrict wheel publish jobs
Only allow requirements to run the wheel publish jobs.  This used to
be set on the build jobs, but we moved them to openstack-zuul-jobs so
they could be speculatively tested.

It makes sense to run the build jobs in the openstack-zuul-jobs gate
when the wheel build scripts are updated.

Thus move the allowed-projects to the publish jobs, which are actually
the bits that only requirements updates should write out.

See also Ica59634eec533db20fe8b1dd148ee9b4d041fd3c

Change-Id: I0574b69090409a6cd3f09f533a4dd7f710f89d2c