Replace build-sphinx-docs jobs
Changes [1] and [2] change the way the build-sphinx-docs jobs work to
not run them on master, use openstack-tox-docs based templates instead.

For networking-dpm, it used the jobs and also the new template. Manual
building shows that tox -e docs works fine without neutron, so remove
the extra build-sphinx-docs jobs.

For valence, use also a template for releasenotes.

[1] https://review.opendev.org/735923
[2] https://review.opendev.org/735937

Change-Id: I930795fd194a14a9094925061608a3764492d55f