gerrit: change retired.config acls
This patch updates the retired.config ACL to allow for the technical
commitee to be able to push changes into the repositories which are
retired.

The ACLs allows tech-committee group members to set all labels onto
changes as well as allowing them exclusive rights to push (therefore not
allowing any other members) and giving them access to submit changes (in
order to skip our gating).

The goal is to evenutally replace this group by another one once the
ACLs are verified to be working.

Change-Id: Ia6d516621ec405b02f3f97340d96d9938b605d8f