Retire networking-l2gw and networking-l2gw-tempest-plugin
networking-l2gw has been moved from neutron stadium projects
in ocata and also marked as retired
- https://review.opendev.org/#/c/392010/2

and networking-l2gw-tempest-plugin was never a official project.

In disucssion with neutron team on IRC, there is no plan or active
team for this projects under neutron stadum so it is good to complete
the retirement of those repo.

- http://eavesdrop.openstack.org/irclogs/%23openstack-tc/%23openstack-tc.2020-06-24.log.html#t2020-06-24T15:31:31

This commit adds the noop job got their gate to start the repo cleanup

Needed-By: https://review.opendev.org/#/c/738038/
Needed-By: https://review.opendev.org/#/c/738035/
Change-Id: Iaecc9d4fa39b388f9f88729ffdd89bc5e431bc9d