Add notifications for openstack-stable channel
It's sometimes hard to be aware of pending stable backports. The stable
team is very small, but they are still around and would like to help
review backports. Since the openstack-stable channel is very quiet
anyway, this adds notifications for new proposed backports to the
channel to help raise awareness of new things to review.

Change-Id: Id29d9d918dbff4777ea4ffa07f1c8441f4c52b8d
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>