Re-add vexxhost-specific labels
These labels were not migrated when we moved to running inside our new
hardware inside the Montreal datacenter.

Those largely should disappear as soon as we move our operator
development to be multi-node to avoid having a 'unicorn' flavor.  We'd
like to use them temporarily to unblock our progress for now.

Change-Id: Ie29a071b1b8916e0d09990b97bf589bf2b0fc50a