Gerritbot: only comment on stable:follows-policy repos
Notifying the #openstack-stable channel of any new stable branch patches
has been useful, but it does produce a lot of noise from repos that do
not assert the stable:follows-policy tag. The stable team only has
rights to these tagged repos, so this refines the list of repos to
notify to only include the ones that assert the tag.

Change-Id: Iad325c74bfd8dcf5657da290dedca779de41f6fa
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>