New Project Request: hostconfig-operator
This is a request to create the airship/hostconfig-operator project in opendev.
Hostconfig-operator is based on the Kubernetes ansible operator framework,
which allows declarative intent (represented as custom resources) to drive
Ansible playbooks.

This operator will be used by Airship 2 to drive declarative host-level
management for the nodes that run the managed Kubernetes cluster;
generally for things that are not handled by existing tools like the
Cluster API or KubeADM.  Examples include setting permissions on the
filesystem, enforcing OS resource limits, etc.

This project plays a similar role in Airship 2 that Divingbell played
in Airship 1.

Change-Id: I22090408b152b15f5b4f823945b098615ca8c2af