Update branch checkout for octavia-lib DIB element
Octavia has a diskimage-builder element to install octavia-lib. When
creating an amphora image from an Octavia stable branch, the expectation
is the octavia-lib code will match the same branch (master or stable).

This patch updates the branch name and upper constraints for octavia-lib
in Octavia on stable branch creation.

This change is in line with I8eba64c886c187c8652f94735ca6153702203d17.

Depends-On: https://review.opendev.org/#/c/745506
Change-Id: I31b762f631304636bafabec9c54cd31c6c91f124