Temporarily disable rax-dfw for mirror reboot
While attempting to replace the Cinder volume attached to the
rax-dfw mirror server, the xenwatch reported a page allocation
failure. In order to ensure the new block device is correctly seen
by the guest OS we should reboot it. Temporarily wind down testing
there so as not to disrupt any running jobs when we do so. This can
be reverted after we complete the reboot.

Change-Id: I96f69bfa7e3bb2928f24755fd1ff2cb1b347548c