Pause image builds on nb04.opendev.org
We'd like to delete nb04.opendev.org and go back to running two x86
image builders. In order to do that gracefully we need to pause all
image builds on nb04 so that nb01 and nb02 take up the slack. Then once
nb04's images are no longer in use they can be deleted and the builder
can be removed.

Change-Id: Ibe00252de4f602a6c9f48ad1799d844710ddc6dc