Remove openstack-python-jobs from GBP templates
The openstack-python-jobs causes problems for stable/ussuri, as it
adds py27 jobs. Since python2.7 is deprecated in ussuri, we need
to remove this from the GBP gate jobs. Once merged, the GBP projects
will need to modify their local zuul configuration to ensure that the
required jobs (e.g. py36, py37) are included.

Change-Id: I07bfb3d20cd7ed855f38de0345f275df06ab0a10