Create inspur/venus-specs repository
Inspur is working on a log collection/management module for OpenStack.
It needs to be released under an open source license and fleshed out
before it can be officially proposed as an offical OpenStack project.

Create a inspur/venus-specs repository and use the same ACL with Venus
project(c187f810289c0567fb2a05a8012f7227d94d0de7) to record the
specifications of the venus preliminary changes.

Change-Id: I9129f2d71cfa211d4d65b5776d3afd02742825e1