Run tag-releases on ubuntu-focal
The tag-releases job installs a distro package of git-review to use
for interacting with Gerrit, but since our Gerrit upgrade to 3.2 the
git-review 1.26 in Ubuntu Bionic is too old to work with it
correctly. Switch this job to run on ubuntu-focal nodes where it
will get a new enough git-review 1.27 instead.

Change-Id: Icb5042525bc0f2ddfaa8be24b3b3ed0e5fb2c625