Update neutron grafana dashboard
Recently we did some change in names of the CI jobs and we moved some
jobs to the experimental and periodic queues.
This patch reflects those changes in the Grafana dashboard.

It also removes graph with number of periodic jobs run every day as
those jobs are run once per day so there is no need to have this graph
at all.

Change-Id: I6984cbd220ec76eec04d0c096c5edf0d2f53f776