Switch to using v3-standard-8 flavors
VEXXHOST no longer offers to flavors of instances, and it makes more
sense to use the standard flavors.  This flavor has 8 cores, 32GB
of memory and 160GB of disk space.  It doesn't make sense to have
a smaller memory/disk allocation as the hypervisors are ratio'd out
to work with this configuration.

Change-Id: Ie424e7d6236c00678d92b34ed33dafaf966180fa