New Project Request: airship/gerrit-to-github-bot
This is a request to create the airship/gerrit-to-github-bot
project in opendev.

This bot synchronizes state between gerrit patchsets, where
airship work is done, and github issues, where airship scope
is tracked.

If you can seed the new gerrit group with me then I can add
the appropriate core team.

The upstream imported project, while created for Airship in
the first place, is essentially abandoned by its owner (hence our
need to import to airship proper to continue enhancing it).
We are unable to clean up the extra feature branch outside of master.

Change-Id: I9c75e12d6b9c2b1c7c4a3047a90e9f80440c3532