Add custom cirros image with ahci module enabled to cache
As discussed on the ML [1] the nova-next job is looking to start testing
the q35 machine type. In order to do this *before* the next Cirros
release a custom dev build of the Cirros image has been built with the
ahci module included, as is now required to allow for SATA based config
drives to work.

[1] http://lists.openstack.org/pipermail/openstack-discuss/2021-March/020823.html

Change-Id: I67912064487598c0e5b4ce3001276f42e0ebcad1