Startup nl01.opendev.org idle
This removes min ready settings for labels on nl01.opendev.org to
prevent it from generating extra min ready requests. We also set
max-servers to 0 on its providers so that it essentailly idles.

This is done to ensure the new server can be deployed properly. A
followon change will flip nl01.openstack.org to idle and set min-ready
and max-servers values on nl01.opendev.org to put it into use.

The old server can be removed once it idles.

Change-Id: Icb13ce9153e3d627b187ec8722539f20db51e266