Add idle configs for nl02-04.opendev.org
These configs will allow us to bring up the nodepool-launcher service on
new focal nodepool launcher nodes in an idle state. Once we are happy
with them in the idle state we will land the next change which flips the
old hosts to idle and puts these new servers in production.

Change-Id: I0d0a30c49c1336b33dd969b039544e10d98dad0e