Stop linting zuul-jobs from here
When we run ansible-lint on project-config the tool is finding and
reporting errors in zuul-jobs. Zuul-jobs appears to be included to
ensure that all the roles we use are defined, but we don't want to lint
the content of those roles.

To address this we add ../../zuul to the ansible lint exclude path which
includes zuul-jobs.

Change-Id: Ie9aeca81ba4fb95df0e7e1ee275230378e49b741